+++ src/libc/include/stdio.h
@@ -6,7 +6,7 @@
 #define NULL ((void*)0)
 #endif
 
-typedef unsigned int size_t;
+typedef long unsigned int size_t;
 
 #define EOF (-1)
